Created MVC architecture, adjusted imports
