+++ server.js
@@ -1,0 +1,6 @@
+const app = require("./app");
+const port = process.env.PORT || 3000;
+
+app.listen(port, () => {
+  console.log(`Server running on port http://localhost:${port}`);
+});
